Move job post queries from view to the model
